move reco-dependent exclusions into millipede reco. Pull omgeo and pulseseries from GCDQp frame
